requirements.txt
pandas
numpy
python-dotenv
certifi
pre-commit
pymongo
scikit-learn
dill
pyaml
-e .
